Organised everything and renamed classes to understand what they do
